requirements.txt
PyQt6
pynput
pyobjc-framework-Cocoa
numpy
